image product, customer, order

+++ application/views/backend/categorys/list.php
@@ -32,7 +32,7 @@
                                         <a class="btn btn-warning d-block" href="<?=base_url()?>backend/edit_category/<?=$value['id']?>">
                                             <i class="align-middle text-white" data-feather="edit"></i>
                                         </a>
-                                        <a class="btn btn-danger d-block" href="<?=base_url()?>control_category/delete_category/<?=$value['id']?>">
+                                        <a class="btn btn-danger d-block" href="<?=base_url()?>control_category/delete_category/<?=$value['id']?>" onclick="return confirm('Are you sure?')">
                                             <i class="align-middle text-white" data-feather="delete"></i>
                                         </a>
                                     </div>
